js: remove node things from non-node modules
