+++ js/src/model-config-node.ts
@@ -1,0 +1,27 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import * as fs from "fs/promises";
+import { ModelConfig } from "./model-config.js";
+
+export class ModelConfigNode extends ModelConfig {
+  async loadFile(configPath: string): Promise<void> {
+    if (this.loaded) {
+      return;
+    }
+    const config = JSON.parse((await fs.readFile(configPath)).toString());
+    this.setConfig(config);
+    this.loaded = true;
+  }
+}
